fix: move functions to header and remove the beratungshilfe.pdf.ts file
