--- app/services/pdf/beratungshilfe/beratungshilfe.pdf.ts
@@ -1,62 +1,0 @@
-import type { BeratungshilfeFormularContext } from "~/models/flows/beratungshilfeFormular";
-import { checkboxListToString } from "../checkboxListToString";
-
-export const staatlicheLeistungMapping = {
-  grundsicherung: "Grundsicherung",
-  asylbewerberleistungen: "Asylbewerberleistungen",
-  buergergeld: "Bürgergeld",
-  andereLeistung: "Andere Leistung",
-  keine: "Keine",
-};
-
-export const getOccupationDetails = (
-  context: BeratungshilfeFormularContext,
-  withAdditionalIncome = true,
-) => {
-  const description: string[] = [];
-
-  if (context.erwerbstaetig === "no") {
-    description.push("nicht erwerbstätig");
-  } else if (context.berufart) {
-    const occupation = "Erwerbstätig";
-    const occupationTypeSelected = checkboxListToString(
-      {
-        selbststaendig: "selbstständig",
-        festangestellt: "festangestellt",
-      },
-      context.berufart,
-    );
-
-    description.push(
-      `${occupation}${
-        occupationTypeSelected ? " (" + occupationTypeSelected + ")" : ""
-      }`,
-    );
-  }
-
-  const berufsituationMapping = {
-    no: "",
-    pupil: "Schüler:in",
-    student: "Student:in",
-    retiree: "Rentner:in",
-  };
-
-  description.push(berufsituationMapping[context.berufsituation ?? "no"]);
-
-  if (context.weitereseinkommen && withAdditionalIncome) {
-    const otherIncomes = checkboxListToString(
-      {
-        unterhaltszahlungen: "Unterhaltszahlungen",
-        wohngeld: "Wohngeld",
-        kindergeld: "Kindergeld",
-        bafoeg: "Bafög",
-        others: "Sonstiges",
-      },
-      context.weitereseinkommen,
-    );
-
-    description.push(otherIncomes);
-  }
-
-  return description.filter((value) => value).join(", ");
-};
