+++ app/services/pdf/beratungshilfe/sections/header.ts
@@ -1,9 +1,5 @@
 import type { BeratungshilfeFormularContext } from "~/models/flows/beratungshilfeFormular";
 import type { BeratungshilfePDF } from "data/pdf/beratungshilfe/beratungshilfe.generated";
-import {
-  getOccupationDetails,
-  staatlicheLeistungMapping,
-} from "../beratungshilfe.pdf";
 import { checkboxListToString } from "../../checkboxListToString";
 import { newPageHint, type Attachment } from "../attachment";
 
@@ -55,3 +51,63 @@ export default function fillHeader(
     pdfFields.berufErwerbstaetigkeit.value = newPageHint;
   }
 }
+
+const staatlicheLeistungMapping = {
+  grundsicherung: "Grundsicherung",
+  asylbewerberleistungen: "Asylbewerberleistungen",
+  buergergeld: "Bürgergeld",
+  andereLeistung: "Andere Leistung",
+  keine: "Keine",
+};
+
+const getOccupationDetails = (
+  context: BeratungshilfeFormularContext,
+  withAdditionalIncome = true,
+) => {
+  const description: string[] = [];
+
+  if (context.erwerbstaetig === "no") {
+    description.push("nicht erwerbstätig");
+  } else if (context.berufart) {
+    const occupation = "Erwerbstätig";
+    const occupationTypeSelected = checkboxListToString(
+      {
+        selbststaendig: "selbstständig",
+        festangestellt: "festangestellt",
+      },
+      context.berufart,
+    );
+
+    description.push(
+      `${occupation}${
+        occupationTypeSelected ? " (" + occupationTypeSelected + ")" : ""
+      }`,
+    );
+  }
+
+  const berufsituationMapping = {
+    no: "",
+    pupil: "Schüler:in",
+    student: "Student:in",
+    retiree: "Rentner:in",
+  };
+
+  description.push(berufsituationMapping[context.berufsituation ?? "no"]);
+
+  if (context.weitereseinkommen && withAdditionalIncome) {
+    const otherIncomes = checkboxListToString(
+      {
+        unterhaltszahlungen: "Unterhaltszahlungen",
+        wohngeld: "Wohngeld",
+        kindergeld: "Kindergeld",
+        bafoeg: "Bafög",
+        others: "Sonstiges",
+      },
+      context.weitereseinkommen,
+    );
+
+    description.push(otherIncomes);
+  }
+
+  return description.filter((value) => value).join(", ");
+};
